requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.1.1
Flask-Login==0.6.2